[IPLUG/EXAMPLES] improved build-mac script

+++ IPlugExamples/buildall-mac.command
@@ -6,27 +6,30 @@
 
 BASEDIR=$(dirname $0)
 
+#here you can choose a target to build
+TARGET="All"
+#TARGET="APP"
+#TARGET="VST2"
+#TARGET="VST3"
+#TARGET="AU"
+#TARGET="RTAS"
+#TARGET="AAX"
+
 cd $BASEDIR
 
 echo "building all example plugins..."
 
+if [ -f build_errors.log ]
+then
+  rm build_errors.log
+fi
+
 for file in * 
 do
-#echo $file
-	if [ -d "$file/$file.xcodeproj" ]
-	then
-		echo "building $file/$file.xcodeproj All targets"
-		xcodebuild -project "$file/$file.xcodeproj" -target "All" -configuration Release 2> ./build_errors.log
-# 		echo "building $file/$file.xcodeproj VST"
-# 		xcodebuild -project "$file/$file.xcodeproj" -target "VST_32&64_intel" -configuration Release
-# 		echo "building $file/$file.xcodeproj AU"
-# 		xcodebuild -project "$file/$file.xcodeproj" -target "AU_32&64_intel" -configuration Release
-# 		echo "building $file/$file.xcodeproj App"
-# 		xcodebuild -project "$file/$file.xcodeproj" -target "OSXAPP_32&64_intel" -configuration Release
-# 		echo "building $file/$file.xcodeproj VST3"
-# 		xcodebuild -project "$file/$file.xcodeproj" -target "VST3_32&64_intel" -configuration Release
-# 		echo "building $file/$file.xcodeproj RTAS"
-# 		xcodebuild -project "$file/$file.xcodeproj" -target "RTAS_32_intel" -configuration Release
+  if [ -d "$file/$file.xcodeproj" ]
+  then
+    echo "building $file/$file.xcodeproj $TARGET target"
+    xcodebuild -project "$file/$file.xcodeproj" -target $TARGET -configuration Release 2> ./build_errors.log
 
     if [ -s build_errors.log ]
     then
@@ -35,10 +38,13 @@ do
       cat build_errors.log
       exit 1
     else
-     rm  build_errors.log
+      if [ -f build_errors.log ]
+      then
+        rm build_errors.log
+      fi
     fi
 
-	fi
+  fi
 done
